@@ -1,6 +1,6 @@
 module github.com/yokomotod/yuccadb
 
-go 1.20
+go 1.22
 
 require (
 	cloud.google.com/go v0.110.2 // indirect
