@@ -3,6 +3,7 @@ package yuccadb_test
 import (
 	"context"
 	"fmt"
+	"log"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -12,10 +13,16 @@ import (
 	"github.com/yokomotod/yuccadb"
 )
 
+const (
+	dataDir       = "./testdata"
+	testFileDir   = "./testfile"
+	testTableName = "test"
+)
+
 func TestMain(m *testing.M) {
-	err := genTestCsv(testFileName())
+	err := genTestCsv()
 	if err != nil {
-		fmt.Println(err)
+		log.Println(err)
 		os.Exit(1)
 	}
 
@@ -25,39 +32,49 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-const size = 1_000_000
+const tableSize = 1_000_000
 
 func testFileName() string {
-	s, unit := size, ""
+	size, unit := tableSize, ""
 	units := []string{"k", "m", "g", "t"}
-	for i := 0; s >= 1_000; s, i = s/1_000, i+1 {
+
+	for i := 0; size >= 1_000; size, i = size/1_000, i+1 {
 		unit = units[i]
 	}
-	return fmt.Sprintf("./testfile/test%d%s.csv", s, unit)
 
+	return fmt.Sprintf("./%s/test%d%s.csv", testFileDir, size, unit)
 }
 
-func genTestCsv(testFile string) error {
+func genTestCsv() error {
+	testFile := testFileName()
+
 	// check test file exists and skip generating
 	if _, err := os.Stat(testFile); err == nil {
-		fmt.Printf("Skip generating %s\n", testFile)
+		log.Printf("Skip generating %s\n", testFile)
+
 		return nil
 	}
 
-	fmt.Printf("Generating %s...\n", testFile)
+	log.Printf("Generating %s...\n", testFile)
 
-	f, err := os.OpenFile(testFile, os.O_WRONLY|os.O_CREATE, 0644)
-	if err != nil {
-		return fmt.Errorf("failed to open file: %s", err)
+	if _, err := os.Stat(testFileDir); os.IsNotExist(err) {
+		if err := os.Mkdir(testFileDir, 0o755); err != nil {
+			return fmt.Errorf("failed to create data directory: %w", err)
+		}
 	}
 
-	for i := 0; i < size; i++ {
+	file, err := os.OpenFile(testFile, os.O_WRONLY|os.O_CREATE, 0o600)
+	if err != nil {
+		return fmt.Errorf("failed to open file: %w", err)
+	}
+
+	for i := range tableSize {
 		key := fmt.Sprintf("%010d", i)
 		value := strconv.Itoa(i)
 
-		_, err := f.WriteString(key + "," + value + "\n")
+		_, err := file.WriteString(key + "," + value + "\n")
 		if err != nil {
-			return fmt.Errorf("failed to write file: %s", err)
+			return fmt.Errorf("failed to write file: %w", err)
 		}
 	}
 
@@ -65,26 +82,30 @@ func genTestCsv(testFile string) error {
 }
 
 func TestDB(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
-	dataDir, testTableName, testFile := "./testdata", "test", testFileName()
+	testFile := testFileName()
 
 	// (re-)create data dir
 	if err := os.RemoveAll(dataDir); err != nil {
 		t.Fatal(err)
 	}
-	if err := os.MkdirAll(dataDir, 0755); err != nil {
+
+	if err := os.MkdirAll(dataDir, 0o755); err != nil {
 		t.Fatal(err)
 	}
 
-	db, err := yuccadb.NewYuccaDB(ctx, dataDir)
+	db, err := yuccadb.NewYuccaDB(dataDir)
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	if err := db.PutTable(ctx, testTableName, testFile, false); err != nil {
 		t.Fatal(err)
 	}
 
-	db2, err := yuccadb.NewYuccaDB(ctx, dataDir)
+	db2, err := yuccadb.NewYuccaDB(dataDir)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -98,8 +119,8 @@ func TestDB(t *testing.T) {
 	}{
 		{"key exists on index", db, "0000000000", "0", true},
 		{"key does not exist on index", db, "0000099999", "99999", true},
-		{"last key", db, fmt.Sprintf("%010d", size-1), strconv.Itoa(size - 1), true},
-		{"before last key", db, fmt.Sprintf("%010d", size-2), strconv.Itoa(size - 2), true},
+		{"last key", db, fmt.Sprintf("%010d", tableSize-1), strconv.Itoa(tableSize - 1), true},
+		{"before last key", db, fmt.Sprintf("%010d", tableSize-2), strconv.Itoa(tableSize - 2), true},
 		{"not found but middle of keys", db, "0000099999x", "", false},
 		{"key exists on reloaded index", db2, "0000000000", "0", true},
 	}
@@ -107,26 +128,37 @@ func TestDB(t *testing.T) {
 	for _, c := range cases {
 		// sub test
 		t.Run(c.name, func(t *testing.T) {
+			t.Parallel()
 
-			res, err := c.db.GetValue(testTableName, c.key)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if !res.TableExists {
-				t.Fatalf("table %s does not exist", testTableName)
-			}
-
-			if res.KeyExists != c.wantKeyExists {
-				t.Fatalf("expected keyExists %t, but got %t", c.wantKeyExists, res.KeyExists)
-			}
-			if res.Value != c.want {
-				t.Fatalf("expected %s, but got %s", c.want, res.Value)
-			}
+			testDBCase(t, c.db, testTableName, c.key, c.want, c.wantKeyExists)
 		})
 	}
 }
 
+func testDBCase(t *testing.T, db *yuccadb.YuccaDB, tableName, key, want string, wantKeyExists bool) {
+	t.Helper()
+
+	res, err := db.GetValue(tableName, key)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if !res.TableExists {
+		t.Fatalf("table %s does not exist", tableName)
+	}
+
+	if res.KeyExists != wantKeyExists {
+		t.Fatalf("expected keyExists %t, but got %t", wantKeyExists, res.KeyExists)
+	}
+
+	if res.Value != want {
+		t.Fatalf("expected %s, but got %s", want, res.Value)
+	}
+}
+
 func TestLoadError(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 
 	tempDir := t.TempDir()
@@ -139,18 +171,20 @@ func TestLoadError(t *testing.T) {
 	content := strings.Join(lines, "\n")
 	brokenFile := filepath.Join(tempDir, "broken.csv")
 
-	if err := os.WriteFile(brokenFile, []byte(content), 0644); err != nil {
+	if err := os.WriteFile(brokenFile, []byte(content), 0o600); err != nil {
 		t.Fatal(err)
 	}
 
-	db, err := yuccadb.NewYuccaDB(ctx, tempDataDir)
+	db, err := yuccadb.NewYuccaDB(tempDataDir)
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	err = db.PutTable(ctx, "broken", brokenFile, false)
 	if err == nil {
 		t.Fatal("expected error")
 	}
+
 	expectedErr := " invalid line:"
 	if !strings.Contains(err.Error(), expectedErr) {
 		t.Fatalf("expected %q to include %q", err.Error(), expectedErr)
@@ -161,16 +195,20 @@ func TestLoadError(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	if len(files) != 0 {
 		filenames := make([]string, len(files))
 		for i, f := range files {
 			filenames[i] = f.Name()
 		}
+
 		t.Fatalf("expected 0 files, but found %s", filenames)
 	}
 }
 
 func TestDuplicateTableError(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 
 	tempDir := t.TempDir()
@@ -178,14 +216,16 @@ func TestDuplicateTableError(t *testing.T) {
 
 	content := "key,value"
 	testFile := filepath.Join(tempDir, "test.csv")
-	if err := os.WriteFile(testFile, []byte(content), 0644); err != nil {
+
+	if err := os.WriteFile(testFile, []byte(content), 0o600); err != nil {
 		t.Fatal(err)
 	}
 
-	db, err := yuccadb.NewYuccaDB(ctx, tempDataDir)
+	db, err := yuccadb.NewYuccaDB(tempDataDir)
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	if err := db.PutTable(ctx, "test", testFile, false); err != nil {
 		t.Fatal(err)
 	}
@@ -194,6 +234,7 @@ func TestDuplicateTableError(t *testing.T) {
 	if err == nil {
 		t.Fatal("expected error")
 	}
+
 	expectedErr := "table test already exists and replace is false"
 	if err.Error() != expectedErr {
 		t.Fatalf("expected error %s, but got %s", expectedErr, err.Error())
@@ -201,46 +242,53 @@ func TestDuplicateTableError(t *testing.T) {
 }
 
 func TestReplaceTable(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 	tableName := "test"
 	tempDir := t.TempDir()
 	tempDataDir := t.TempDir()
 	testFile := filepath.Join(tempDir, "test.csv")
 
-	db, err := yuccadb.NewYuccaDB(ctx, tempDataDir)
+	db, err := yuccadb.NewYuccaDB(tempDataDir)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	content := "key,value"
-	if err := os.WriteFile(testFile, []byte(content), 0644); err != nil {
+	if err := os.WriteFile(testFile, []byte(content), 0o600); err != nil {
 		t.Fatal(err)
 	}
+
 	if err := db.PutTable(ctx, tableName, testFile, false); err != nil {
 		t.Fatal(err)
 	}
+
 	res, err := db.GetValue(tableName, "key")
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	if res.Value != "value" {
 		t.Fatalf("expected value, but got %s", res.Value)
 	}
 
 	// replace
 	content = "key,value2"
-	if err := os.WriteFile(testFile, []byte(content), 0644); err != nil {
+	if err := os.WriteFile(testFile, []byte(content), 0o600); err != nil {
 		t.Fatal(err)
 	}
+
 	if err := db.PutTable(ctx, tableName, testFile, true); err != nil {
 		t.Fatal(err)
 	}
+
 	res, err = db.GetValue(tableName, "key")
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	if res.Value != "value2" {
 		t.Fatalf("expected value, but got %s", res.Value)
 	}
-
 }
