@@ -3,14 +3,15 @@ package main
 
 import (
 	"bytes"
+	"context"
 	"crypto/tls"
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"math/rand"
 	"net/http"
+	"os"
 	"strconv"
 	"time"
 
@@ -20,26 +21,37 @@ import (
 // This is a tool like Apache Benchmark a.k.a "ab".
 // It doesn't implement all the features supported by ab.
 
-var client *http.Client
-var postBody []byte
+//nolint:gochecknoglobals
+var (
+	client   *http.Client
+	postBody []byte
+)
 
+//nolint:gochecknoglobals
 var verbose bool
 
-var method string
-var url string
-var timeout int
-var postFile string
-var contentType string
+//nolint:gochecknoglobals
+var (
+	method      string
+	url         string
+	timeout     int
+	postFile    string
+	contentType string
+)
 
-var disableCompression bool
-var disableKeepalive bool
+//nolint:gochecknoglobals
+var (
+	disableCompression bool
+	disableKeepalive   bool
+)
 
 const size = 100_000_000
 
 func worker() {
-	key := fmt.Sprintf("%010d", rand.Intn(size))
+	key := fmt.Sprintf("%010d", rand.Intn(size)) //nolint:gosec
 	url := url + key
-	request, err := http.NewRequest(method, url, bytes.NewBuffer(postBody))
+
+	request, err := http.NewRequestWithContext(context.Background(), method, url, bytes.NewBuffer(postBody))
 	if err != nil {
 		log.Fatalf("%v\n", err)
 	}
@@ -50,36 +62,40 @@ func worker() {
 	response, err := client.Do(request)
 	elapsed := time.Since(startTime)
 
-	if err != nil {
+	if err != nil { //nolint:nestif
 		if verbose {
 			log.Printf("%v\n", err)
 		}
+
 		boomer.RecordFailure("http", "error", 0.0, err.Error())
 	} else {
 		boomer.RecordSuccess("http", strconv.Itoa(response.StatusCode),
 			elapsed.Nanoseconds()/int64(time.Millisecond), response.ContentLength)
 
 		if verbose {
-			body, err := ioutil.ReadAll(response.Body)
+			body, err := io.ReadAll(response.Body)
 			if err != nil {
 				log.Printf("%v\n", err)
 			} else {
 				log.Printf("Status Code: %d\n", response.StatusCode)
 				log.Println(string(body))
 			}
-
 		} else {
-			io.Copy(ioutil.Discard, response.Body)
+			_, err = io.Copy(io.Discard, response.Body)
+			if err != nil {
+				log.Fatalf("%v\n", err)
+			}
 		}
 
 		response.Body.Close()
 	}
 }
 
+//nolint:funlen
 func main() {
 	flag.StringVar(&method, "method", "GET", "HTTP method, one of GET, POST")
 	flag.StringVar(&url, "url", "", "URL")
-	flag.IntVar(&timeout, "timeout", 10, "Seconds to max. wait for each response")
+	flag.IntVar(&timeout, "timeout", 10, "Seconds to max. wait for each response") //nolint:gomnd
 	flag.StringVar(&postFile, "post-file", "", "File containing data to POST. Remember also to set --content-type")
 	flag.StringVar(&contentType, "content-type", "text/plain", "Content-type header")
 
@@ -112,19 +128,21 @@ verbose: %t`, method, url, timeout, postFile, contentType, disableCompression, d
 		if postFile == "" {
 			log.Fatalln("--post-file can't be empty string when method is POST")
 		}
-		tmp, err := ioutil.ReadFile(postFile)
+
+		tmp, err := os.ReadFile(postFile)
 		if err != nil {
 			log.Fatalf("%v\n", err)
 		}
+
 		postBody = tmp
 	}
 
-	http.DefaultTransport.(*http.Transport).MaxIdleConnsPerHost = 2000
-	tr := &http.Transport{
+	http.DefaultTransport.(*http.Transport).MaxIdleConnsPerHost = 2000 //nolint:forcetypeassert
+	tr := &http.Transport{                                             //nolint:varnamelen
 		TLSClientConfig: &tls.Config{
-			InsecureSkipVerify: true,
+			InsecureSkipVerify: true, //nolint:gosec
 		},
-		MaxIdleConnsPerHost: 2000,
+		MaxIdleConnsPerHost: 2000, //nolint:gomnd
 		DisableCompression:  disableCompression,
 		DisableKeepAlives:   disableKeepalive,
 	}
@@ -135,7 +153,7 @@ verbose: %t`, method, url, timeout, postFile, contentType, disableCompression, d
 
 	task := &boomer.Task{
 		Name:   "worker",
-		Weight: 10,
+		Weight: 10, //nolint:gomnd
 		Fn:     worker,
 	}
 
